Landform Classification after Jasiewicz and Stepinkski (2013)

from https://github.com/m-saenger/geomorphon/blob/master/geomorphon.R

geomorph_ras(
  ras,
  outfile = NULL,
  windowSize = 11,
  flatnessThresh = 1,
  doNew = TRUE
)

Arguments

ras

input DTM (RasterLayer class).

outfile

Name of output file. This string will have '_geomorph' appended. Passed to writeRaster.

windowSize

size (in terms of cells per side) of the neighborhood (moving window) to be used; it must be an odd integer.

flatnessThresh

If slope < flatnessThresh then flat.

doNew

Logical. If outfile already exists, redo anyway.

Value

raster.

Details

M. Sänger 10/2019

Raster data to be on equally spaced grid (e.g. km) Output very sensitive to flatness threshold Function not optmised for performance - run time very long for large data sets