Create dataframe of 'cells' and their associated env data

Multicore raster::extract() adapted from the Stack Exchange Network post by thiagoveloso.

make_env(x, df, cores = 1, out_file = tempfile(), limit = TRUE)

Arguments

x
Raster* object.

df
Dataframe. Must have a column called 'cell' that corresponds to the cell numbers in x.

cores
Numeric. Number of cores to use in snowfall::sfInit().

out_file
Character. Path to save outputs.

limit
Logical. If TRUE, only cells in df will be returned. Can be useful if cells already in out_file have been orphaned.

Value

Dataframe with column cell plus columns equal to the length of x.

Details

This function has been surpassed by get_env_data which uses terra::extract.