Use meta data to name raster cube paths/files

Either create file path for saving a cube or parse the meta data from a cube

name_env_tif(
  x,
  dir_only = FALSE,
  skips = c("base"),
  prefixes = c("sat", "use"),
  file_type = c("\\.tif", "\\.nc"),
  parse = FALSE,
  fill_null = FALSE,
  x_null = 3,
  ...
)

Arguments

x

Either dataframe with path(s) to parse (in column path), character vector of path(s) to search, or named list object.

dir_only

Logical.

skips

Character. When parsing, skip (filter) any files with a match to skips

prefixes

Character. If x is a named list, any name prefixes to remove before matching to naming structure.

file_type

Character. e.g. "tif". More than one can be provided but if parse then file_type[1] is used for out_file file type.

parse

Logical. If FALSE an out_file path will be returned rather than parsed. Assumes the appropriately names columns are supplied in df.

fill_null

Logical. If TRUE, will fill up to x_null definitions with NULL (and issue a warning).

x_null

Numeric. Even if fill_null is TRUE, if there are more than x_null missing definitions, an error will be thrown.

Value

If !parse, tibble with columns:

vector

Vector layer originally used to define area of interest

filt_col

Column name from vector to filter to define area of interest

level

Level(s) of filt_col originally filtered to define area of interest

buffer

Any buffer around area of interest

period

Period represented within layer using ISO8601-compliant time period for regular data cubes. e.g., "P16D" for 16 days

res

Resolution of output raster in units of rasters crs

source

e.g. DEA is digital earth Austrlia

collection

e.g. ga_ls8c_ard_3 is landsat 8 collection in DEA

layer

Band or layer name

start_date

Start date for layer

type

File type of path.

path

Full (relative) path including file_type.

If !parse, df with additional column out_file