Prepare a tibble of environmental layers from settings

'Seasons' functionality removed 2025-09-05. If seasons are needed, left join the result of prepare_env() to the result of make_seasons()``$seasons.

prepare_env(
  set_list,
  base_dir = if (Sys.info()["sysname"] == "Windows") "I:" else fs::path("/mnt/envcube",
    ""),
  create_short_desc = TRUE,
  max_date_only = TRUE,
  exclude_list = NULL,
  ...
)

Arguments

set_list

list. usually from yaml::read_yaml("settings/setup.yaml")

base_dir

Character. Passed to base_dir argument of envFunc::name_env_out()

create_short_desc

Logical. Create a short description to use as a unique 'name'. Requires columns 'name' and 'func'.

max_date_only

Logical. Filter rows duplicated by layer and func to the latest start_date.

exclude_list

Named list used to exclude rasters, specifying the column name and levels (used as regex) to exclude (filter) within that column.

...

Passed to envRaster::name_env_tif

Value

tibble of raster paths and meta data parsed into columns