Summarise raster files

Rasters are stacked via terra::rast(), then summaries are generated by terra::app(). Summaries are saved to file. Rasters are split by index (using terra::subset()) before summarising, allowing for multi-layer rasters. Lyrn is added to filename.

summarise_rast_paths(
  paths,
  process,
  season = NA,
  epoch = NA,
  layer = 1L,
  out_dir,
  force_new = FALSE,
  methods = c("mean", "median", "min", "max", "sd"),
  out_type = "tif",
  ...
)

Arguments

paths
Character. Paths to any raster that can be read by terra::rast().

process, season, epoch
Character. Descriptors for rasters being summarised. See envEcosystems::env.

layer
Numeric. Layer/band of paths to summarise. Only one band at a time can be summarised.

out_dir
Character. Path for saving summarised rasters.

force_new
Logical. Run summarise even if resulting file already exists.

methods
Method(s) to apply to each vector of stacked cells. Name of function(s).

out_type
Extension for raster file written. Required as filename is created from out_base and method so can't be passed directly.

...
Passed to terra::app(). Thus, (presumably) includes options to terra::writeRaster() via wopt argument (untested).

Value

Side effect of saving summarised raster(s) to file.

Examples