Widen env data, especially AusCover data from get_env_data.

Output is a context by environmental variable data frame.

widen_env_data(
  env_data_long,
  orig_df,
  context,
  env_groups = c("process", "layer", "method", "season"),
  epoch_step = 10,
  min_years = 3
)

Arguments

env_data_long
Dataframe, in long format. Usually resulting from get_env_data.

orig_df
Dataframe containing the locations passed to get_env_data.

context
Character. Columns in orig_df defining the context.

env_groups
Character. Name of columns in env_data_long that define the context of the raster. Default to 'process', 'layer', 'method' and 'season'. See envEcosystems::env for definitions.

epoch_step
Numeric. If summarising several (years) rasters to one for the given context, how many years to summarise?

min_years
Numeric. What is the minimum acceptable number of years with data. Points with less than min_years of data will be filtered.

Value

Site (context) by environmental data with one row of env data for each context.