Title R function for landform classification on the basis on Topographic Position Index

from https://rdrr.io/github/gianmarcoalberti/GmAMisc/src/R/landfClass.R. Function seemed to disappear from GmAMisc package between 1.1.0 and 1.1.1

land_class(
  ras,
  outfile = NULL,
  doNew = TRUE,
  sn = 3,
  ln = 7,
  n.classes = "six",
  win = "rectangle"
)

Arguments

ras

input DTM (RasterLayer class).

outfile

character. Optional path filename passed to writeRaster.

doNew

logical. If FALSE, will only process x if outfile does not exist.

sn

size (in terms of cells per side) of the neighborhood (moving window) to be used; it must be an odd integer. s(mall) n(eighbourhood).

ln

if the 10-class classification is selected, this paramenter sets the l(arge) n(eighborhood) to be used.

n.classes

"six" or "ten" for a six- or ten-class landform classification.

win

character. Shape of the moving window used by spatialEco::tpi.

Details

The function allows to perform landform classification on the basis of the Topographic Position Index calculated from an input Digital Terrain Model (RasterLayer class).

The TPI is the difference between the elevation of a given cell and the average elevation of the surrounding cells in a user defined moving window. For landform classification, the TPI is first standardized and then thresholded; to isolate certain classes, a slope raster (which is internally worked out) is also needed.
For details about the implemented classification, see: http://www.jennessent.com/downloads/tpi_documentation_online.pdf.

Two methods are available:

the first (devised by Weiss) produces a 6-class landform classification comprising -- valley
-- lower slope
-- flat slope
-- middle slope
-- upper clope
-- ridge
-the second (devised by Jennes) produces a 10-class classification comprising -- canyons, deeply incised streams
-- midslope drainages, shallow valleys
-- upland drainages, headwaters
-- u-shaped valleys
-- plains
-- open slopes
-- upper slopes, mesas
-- local ridges, hills in valleys
-- midslope ridges, small hills
-- mountain tops, high ridges
The second classification is based on two TPI that make use of two neighborhoods (moving windows) of different size: a s(mall) n(eighborhood) and a l(arge) n(eighborhood), defined by the parameters sn and ln.