Generate lists of paths in preparation for summarise_rast_paths().

Using the file naming conventions in the AusCover data , collate paths of previously downloaded raster files in preparation for running envImport::summarise_rast_paths().

prep_auscover(
  dir_local = "../../data/raster/dynamic/AusCover/landsat/",
  epoch_step = 10,
  epoch_overlap = FALSE,
  epoch_which = -1,
  filt_n = 9
)

Arguments

dir_local

Character. Path to search for rasters to summarise.

epoch_step

Numeric. How many years in an epoch?

epoch_overlap

Logical. Should epochs overlap by one year? i.e. epoch_overlap = TRUE gives, say, 2000-2010 and 2010-2020 whereas epoch_overalp = FALSE gives, say, 2000-2009 and 2010-2019.

epoch_which

Numeric. Which epochs to summarise? Zero for current epoch, -1 for immediate previous epoch etc. -2:-1 for the last two full epochs.

filt_n

Numeric. Filter prepared stacks with less than filt_n rasters.

Value

Dataframe with columns

process

Raster product in data. See envEcosystems::env.

epoch

Last two digits of initial year and final year for rasters in data. e.g. 10-19.

season

Season for rasters in data.

data

List column of full path of raster files to summarise.